+++ test/src/features/places/place_screen_test.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:golden_toolkit/golden_toolkit.dart';
+import 'package:where_to_go_today/src/features/places/place_screen.dart';
+import 'package:where_to_go_today/src/localization/l10n.dart';
+import 'package:where_to_go_today/src/res/theme/app_theme.dart';
+
+void main() {
+  testGoldens(
+    'Тест на вёрстку экрана PlaceScreen',
+    (tester) async {
+      await tester.pumpWidgetBuilder(
+        Localizations(
+          delegates: const [
+            AppLocalizations.delegate,
+            GlobalMaterialLocalizations.delegate,
+            GlobalWidgetsLocalizations.delegate,
+            GlobalCupertinoLocalizations.delegate,
+          ],
+          locale: const Locale('ru'),
+          child: const PlaceScreen(),
+        ),
+        wrapper: materialAppWrapper(theme: AppTheme.lightTheme),
+      );
+
+      await multiScreenGolden(tester, 'place_screen');
+    },
+  );
+}